+++ build.py
@@ -5,7 +5,7 @@
 # Program: build.py
 # Author:  Mauricio Caceres Bravo <mauricio.caceres.bravo@gmail.com>
 # Created: Sun Oct 15 10:26:39 EDT 2017
-# Updated: Wed Feb 02 11:44:49 EST 2022
+# Updated: Tue Mar 15 00:56:27 EDT 2022
 # Purpose: Main build file for gtools (copies contents into ./build and
 #          puts a .zip file in ./releases)
 
@@ -147,6 +147,8 @@ gtools_ssc = [
     "gstats_range.sthlp",
     "gstats_moving.sthlp",
     "gstats_winsor.sthlp",
+    "gstats_residualize.sthlp",
+    "gstats_hdfe.sthlp",
     "gstats_tab.sthlp",
     "gstats_sum.sthlp",
     "gstats_summarize.sthlp",
@@ -340,6 +342,11 @@ copy2(
 )
 
 copy2(
+    path.join("docs", "stata", "gstats_hdfe.sthlp"),
+    path.join("docs", "stata", "gstats_residualize.sthlp")
+)
+
+copy2(
     path.join("docs", "stata", "gquantiles.sthlp"),
     path.join("docs", "stata", "fasterxtile.sthlp")
 )
@@ -369,34 +376,36 @@ copy2(
     path.join("docs", "stata", "gstats_moving.sthlp")
 )
 
-copy2(path.join("docs", "stata", "gcollapse.sthlp"),        gdir)
-copy2(path.join("docs", "stata", "gcontract.sthlp"),        gdir)
-copy2(path.join("docs", "stata", "gegen.sthlp"),            gdir)
-copy2(path.join("docs", "stata", "gunique.sthlp"),          gdir)
-copy2(path.join("docs", "stata", "gdistinct.sthlp"),        gdir)
-copy2(path.join("docs", "stata", "glevelsof.sthlp"),        gdir)
-copy2(path.join("docs", "stata", "gtop.sthlp"),             gdir)
-copy2(path.join("docs", "stata", "gtoplevelsof.sthlp"),     gdir)
-copy2(path.join("docs", "stata", "gisid.sthlp"),            gdir)
-copy2(path.join("docs", "stata", "greshape.sthlp"),         gdir)
-copy2(path.join("docs", "stata", "greg.sthlp"),             gdir)
-copy2(path.join("docs", "stata", "gregress.sthlp"),         gdir)
-copy2(path.join("docs", "stata", "givregress.sthlp"),       gdir)
-copy2(path.join("docs", "stata", "gglm.sthlp"),             gdir)
-copy2(path.join("docs", "stata", "gpoisson.sthlp"),         gdir)
-copy2(path.join("docs", "stata", "gstats.sthlp"),           gdir)
-copy2(path.join("docs", "stata", "gstats_transform.sthlp"), gdir)
-copy2(path.join("docs", "stata", "gstats_range.sthlp"),     gdir)
-copy2(path.join("docs", "stata", "gstats_moving.sthlp"),    gdir)
-copy2(path.join("docs", "stata", "gstats_winsor.sthlp"),    gdir)
-copy2(path.join("docs", "stata", "gstats_summarize.sthlp"), gdir)
-copy2(path.join("docs", "stata", "gstats_sum.sthlp"),       gdir)
-copy2(path.join("docs", "stata", "gstats_tab.sthlp"),       gdir)
-copy2(path.join("docs", "stata", "gduplicates.sthlp"),      gdir)
-copy2(path.join("docs", "stata", "gquantiles.sthlp"),       gdir)
-copy2(path.join("docs", "stata", "fasterxtile.sthlp"),      gdir)
-copy2(path.join("docs", "stata", "hashsort.sthlp"),         gdir)
-copy2(path.join("docs", "stata", "gtools.sthlp"),           gdir)
+copy2(path.join("docs", "stata", "gcollapse.sthlp"),          gdir)
+copy2(path.join("docs", "stata", "gcontract.sthlp"),          gdir)
+copy2(path.join("docs", "stata", "gegen.sthlp"),              gdir)
+copy2(path.join("docs", "stata", "gunique.sthlp"),            gdir)
+copy2(path.join("docs", "stata", "gdistinct.sthlp"),          gdir)
+copy2(path.join("docs", "stata", "glevelsof.sthlp"),          gdir)
+copy2(path.join("docs", "stata", "gtop.sthlp"),               gdir)
+copy2(path.join("docs", "stata", "gtoplevelsof.sthlp"),       gdir)
+copy2(path.join("docs", "stata", "gisid.sthlp"),              gdir)
+copy2(path.join("docs", "stata", "greshape.sthlp"),           gdir)
+copy2(path.join("docs", "stata", "greg.sthlp"),               gdir)
+copy2(path.join("docs", "stata", "gregress.sthlp"),           gdir)
+copy2(path.join("docs", "stata", "givregress.sthlp"),         gdir)
+copy2(path.join("docs", "stata", "gglm.sthlp"),               gdir)
+copy2(path.join("docs", "stata", "gpoisson.sthlp"),           gdir)
+copy2(path.join("docs", "stata", "gstats.sthlp"),             gdir)
+copy2(path.join("docs", "stata", "gstats_transform.sthlp"),   gdir)
+copy2(path.join("docs", "stata", "gstats_range.sthlp"),       gdir)
+copy2(path.join("docs", "stata", "gstats_moving.sthlp"),      gdir)
+copy2(path.join("docs", "stata", "gstats_winsor.sthlp"),      gdir)
+copy2(path.join("docs", "stata", "gstats_residualize.sthlp"), gdir)
+copy2(path.join("docs", "stata", "gstats_hdfe.sthlp"),        gdir)
+copy2(path.join("docs", "stata", "gstats_summarize.sthlp"),   gdir)
+copy2(path.join("docs", "stata", "gstats_sum.sthlp"),         gdir)
+copy2(path.join("docs", "stata", "gstats_tab.sthlp"),         gdir)
+copy2(path.join("docs", "stata", "gduplicates.sthlp"),        gdir)
+copy2(path.join("docs", "stata", "gquantiles.sthlp"),         gdir)
+copy2(path.join("docs", "stata", "fasterxtile.sthlp"),        gdir)
+copy2(path.join("docs", "stata", "hashsort.sthlp"),           gdir)
+copy2(path.join("docs", "stata", "gtools.sthlp"),             gdir)
 
 copy2(path.join("src", "ado", "_gtools_internal.ado"),  gdir)
 copy2(path.join("src", "ado", "_gtools_internal.mata"), gdir)